requirements.txt
autopep8==1.6.0
pycodestyle==2.8.0
pyenchant==3.2.2
toml==0.10.2
